+++ src/components/commentBox/commentBox.scss
@@ -1,0 +1,48 @@
+.comment_box {
+    display : flex;
+    justify-content : space-between;
+    flex-direction : row;
+    align-items : center;
+    flex-wrap : wrap;
+
+    // background-color : #D8D8D8;
+
+    font-size : 15px;
+
+    margin : 10px auto;
+    padding : 7px;
+
+    height : 40px;
+
+    border-radius : 5px;
+
+    .buttomBox {
+        font-size : 15px;
+        font-weight : none; 
+    
+    }
+
+    .editButtom {
+        margin-right : 10px;
+        border : 0px;
+        margin : 5px 5px auto; 
+    }
+
+    .deleteButtom {
+        border : 0px;
+        margin : 5px 5px auto;  
+    }
+}
+
+.comment_writer {
+    font-weight : bold;
+
+    width : 180px;
+    text-align : center;
+
+    margin-left : -50px;
+}
+
+.comment_text {
+    width : 250px;
+}
