+++ src/components/commentList/CommentList.jsx
@@ -1,10 +1,11 @@
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
-import CommentBox from './CommentBox';
-import CommentForm from './CommentForm';
+import CommentBox from '../commentBox/CommentBox';
+import CommentForm from '../commentForm/CommentForm';
 import { useSelector, useDispatch } from 'react-redux';
-import { __getComments } from '../modules/comment';
+import { __getComments } from '../../modules/comment';
+import "./commentList.scss"
 
 
 
@@ -19,7 +20,7 @@ function CommentList() {
       }, [dispatch]);
 
     return (
-        <StCommentList>
+        <div className='comment_list'>
             <CommentForm/>
             {commentList.map((comment) => (
                 // Number(postId.postId) === comment.postId &&
@@ -29,21 +30,8 @@ function CommentList() {
                 />
             ))}
             
-        </StCommentList>
+        </div>
     );
 }
 
 export default CommentList;
-
-const StCommentList = styled.div`
-    display : flex;
-    justify-content : center;
-    flex-direction : row;
-    align-item : center;
-    flex-wrap : wrap;
-
-    background-color : white;
-    font-size : 15px;
-
-    border-radius : 3px;
-`
