+++ src/components/detailBox/detailBox.scss
@@ -1,0 +1,47 @@
+.detail_box {
+    display : flex;
+    justify-content : space-around;
+    flex-direction : column;
+
+    border: 2px;
+
+    margin-top : 10px;
+    margin-bottom : 10px;
+
+    height : 600px;
+}
+
+.post_writer_box {
+    margin-left : 10px;
+    border-bottom : 1px solid #D8D8D8;
+    padding-bottom : 20px;
+}
+.user_img {
+    width : 30px;
+    height : 30px;
+    border : 0.5px solid #E6E6E6;
+    margin-right : 10px;
+}
+
+.img_box {
+    box-shadow: rgba(0, 0, 0, 0.1) 0px 10px 15px -3px, rgba(0, 0, 0, 0.05) 0px 4px 6px -2px;
+    width : 578px;
+    height : 400px;
+}
+
+.deatil_img {
+    display : block;
+    margin : 0px auto;
+
+    width : 567px;
+    height : 398px;
+}
+
+.post_text {
+    border-bottom : 1px solid #D8D8D8;
+    
+    margin : 11px;
+    padding-top : 30px;
+    padding-bottom : 30px;
+
+}
