+++ src/detailPage/DetailPage.jsx
@@ -1,36 +1,21 @@
 import React from 'react';
 
 import Header from '../components/header/Header';
-import CommentList from '../components/CommentList';
-import DetailBox from '../components/DetailBox';
-import styled from 'styled-components';
+import CommentList from '../components/commentList/CommentList';
+import DetailBox from '../components/detailBox/DetailBox';
+import "./detailPage.scss"
 
 function DetailPage() {
     return (
         <div>
             <Header/>
-        <StDetailPage>
+        <div className='detail_page'>
             <DetailBox/>
             <CommentList/>
-        </StDetailPage>
+        </div>
 
         </div>
     );
 }
 
 export default DetailPage;
-const StDetailPage = styled.div`
-    position : absolute;
-    left : 28%;
-
-    display : flex;    
-    flex-direction : column;
-    justify-content : center;
-    
-    width : 600px;
-    max-width : 1200px;
-    border : 1px solid #D8D8D8;
-
-    padding : 10px;
-    margin-top : 20px;
-`
