+++ src/detailPage/detailPage.scss
@@ -1,0 +1,15 @@
+.detail_page {
+    position : absolute;
+    left : 28%;
+
+    display : flex;    
+    flex-direction : column;
+    justify-content : center;
+    
+    width : 600px;
+    max-width : 1200px;
+    border : 1px solid #D8D8D8;
+
+    padding : 10px;
+    margin-top : 20px;
+}
